requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.6
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
gotrue==2.11.3
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
inflect==7.5.0
Jinja2==3.1.5
jiter==0.8.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.59.7
openai-whisper==20240930
packaging==24.2
postgrest==0.19.3
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.3.0
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
whisper==1.1.10
yarl==1.18.3
